malloc uses size_t for the number of bytes to allocate.
git-svn-id: svn+ssh://danse.us/reflpak@337 6e772d39-f918-0410-b64e-d616bd391b2c

+++ src/cdata.c
@@ -11,7 +11,7 @@
 #ifndef MALLOC
 #define MALLOC malloc
 #else
-extern void* MALLOC(int);
+extern void* MALLOC(size_t);
 #endif
 
 void freeCdata(void)
